refactor(session): extract listing helpers
Problem: the session safety adaptation pushed internal/agent/save.go beyond its existing repolint file-size budget.

Root cause: new listing helpers were added to an already baselined large persistence file.

Fix: move only the new mapping, artifact probe, and error-aware preview helpers into a focused package file without changing behavior or lint baselines.

Verification: go run ./tools/repolint; go test ./internal/agent; go test -race ./internal/agent; git diff --check
